refactor: event helpers
