+++ ayon_server/events/eventstream.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from typing import Any, Awaitable, Callable, Type
 
-from ayon_server.exceptions import ConstraintViolationException
+from ayon_server.exceptions import ConstraintViolationException, NotFoundException
 from ayon_server.lib.postgres import Postgres
 from ayon_server.lib.redis import Redis
 from ayon_server.utils import SQLTool, json_dumps
@@ -220,3 +220,30 @@ class EventStream:
             await Redis.publish(json_dumps(message))
             return True
         return False
+
+    @classmethod
+    async def get(cls, event_id: str) -> EventModel:
+        query = "SELECT * FROM events WHERE id = $1", event_id
+        event: EventModel | None = None
+        async for record in Postgres.iterate(*query):
+            event = EventModel(
+                id=record["id"],
+                hash=record["hash"],
+                topic=record["topic"],
+                project=record["project_name"],
+                user=record["user_name"],
+                sender=record["sender"],
+                depends_on=record["depends_on"],
+                status=record["status"],
+                retries=record["retries"],
+                description=record["description"],
+                payload=record["payload"],
+                summary=record["summary"],
+                created_at=record["created_at"],
+                updated_at=record["updated_at"],
+            )
+            break
+
+        if event is None:
+            raise NotFoundException("Event not found")
+        return event
